All warnings. Prepare share-identity confirmation. Proper integer types.

+++ Smoke/app/src/main/java/org/purple/smoke/Database.java
@@ -141,7 +141,8 @@ public class Database extends SQLiteOpenHelper
 	new ReentrantReadWriteLock();
     private final static String DATABASE_NAME = "smoke.db";
     private final static int DATABASE_VERSION = 1;
-    private final static int WRITE_PARTICIPANT_TIME_DELTA = 60000; // 60 Seconds
+    private final static long WRITE_PARTICIPANT_TIME_DELTA =
+	60000; // 60 Seconds
     private static Database s_instance = null;
     public enum ExceptionLevels
     {
